backend/requirements.txt
fastapi>=0.115.12,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
pydantic>=2.11.3,<3.0.0
pydantic-settings>=2.8.1,<3.0.0
python-multipart>=0.0.20,<1.0.0

requests>=2.32.3,<3.0.0
beautifulsoup4>=4.13.3,<5.0.0
pdfplumber>=0.11.5,<1.0.0
python-dotenv>=1.0.1,<2.0.0

easyocr>=1.7.2,<2.0.0
numpy>=2.2.0,<3.0.0
opencv-python-headless>=4.11.0.86,<5.0.0

torch>=2.9.0,<2.12.0
transformers>=4.51.2,<5.0.0
datasets>=3.5.0,<4.0.0
peft>=0.15.2,<1.0.0
trl>=0.17.0,<1.0.0
accelerate>=1.6.0,<2.0.0
bitsandbytes>=0.45.5,<1.0.0; platform_system != "Windows"
sentencepiece>=0.2.0,<1.0.0
huggingface_hub>=0.30.2,<1.0.0
